requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
pypdf
python-dotenv
streamlit
tiktoken
reportlab
arize-phoenix
openinference-instrumentation-langchain
opentelemetry-sdk
opentelemetry-exporter-otlp